+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Add `sdwan_application_priority_feature_profile` resource and data source
 - Add `sdwan_application_priority_qos_policy_profile_parcel` resource and data source
 - Add `sdwan_transport_wan_vpn_interface_ethernet_profile_parcel` resource and data source
+- Add `sdwan_application_priority_traffic_policy_profile_parcel` resource and data source
 
 ## 0.3.9